hv: pgtable: use new algo to calculate PPT/EPT_PD_PAGE_NUM
In order to support Anderson Lake which physical address width bits is 46,
the current code need to reserve 2^16 PD page ((2^46) / (2^30)). This is
a complete waste of memory.

This patch would reserve PD page according to the numbers of 64 bits PCI MMIO
BAR. For a PCI MMIO BAR: (a) MMIO BAR size must be a power of 2 from 16 bytes;
(b) MMIO BAR base address must be power of two in size and are aligned with
its size. As a result, a PCI MMIO BAR which size is not larger than 1G, it only
spends a PD page to do memory mapping.

Signed-off-by: Li Fei1 <fei1.li@intel.com>