hv: tee: reserve the GPA for TEE SMP boot
The REE and TEE are keep identical mapping for now. This cause the AP's
start-up routine address met conflict as it only allowed 1MB below
memory region.

This patch will dive a 4K hole in REE's 1MB region for TEE using. But it
may not work after enable the MEE encrypt memory region. So need
consider the formal solution later.

Signed-off-by: Jie Deng <jie.deng@intel.com>
Reviewed-by: Wang, Yu1 <yu1.wang@intel.com>